Only run configuration related modules if first boot or dev or not config related or config related and incoming config changes.

+++ provisioners/redhat/provision.sh
@@ -75,6 +75,10 @@ else
 fi
 
 
+# determine if there are configuration changes
+
+
+
 # provision server
 # @todo standardize apache/mysql to match server name?
 if [ "${4}" = "apache" ]; then
@@ -84,18 +88,21 @@ if [ "${4}" = "apache" ]; then
 
     cat "/catapult/provisioners/provisioners.yml" | shyaml get-values-0 redhat.servers.redhat.modules |
     while read -r -d $'\0' key value; do
-        start=$(date +%s)
-        echo -e "\n\n\n==> MODULE: ${key}"
-        echo -e "==> DESCRIPTION: $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.description)"
-        bash "/catapult/provisioners/redhat/modules/${key}.sh" $1 $2 $3 $4
-        end=$(date +%s)
-        echo -e "==> MODULE: ${key}"
-        echo -e "==> DURATION: $(($end - $start)) seconds"
+        # first boot || dev || not config related || config related and incoming config changes
+        if ([ -s /catapult/provisioners/redhat/logs/apache.log ]) || ([ $1 == "dev" ]) || ([ $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.configuration) == "False" ]) || ([ $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.configuration) == "True" ] && [ $(cd /catapult && sudo git diff --exit-code ${branch} origin/${branch} "secrets/configuration.yml.gpg") ]); then
+            start=$(date +%s)
+            echo -e "\n\n\n==> MODULE: ${key}"
+            echo -e "==> DESCRIPTION: $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.description)"
+            bash "/catapult/provisioners/redhat/modules/${key}.sh" $1 $2 $3 $4
+            end=$(date +%s)
+            echo -e "==> MODULE: ${key}"
+            echo -e "==> DURATION: $(($end - $start)) seconds"
+        fi
     done
 
     provisionend=$(date +%s)
     echo -e "\n\n\n==> PROVISION: apache"
-    echo -e "==> DURATION: $(($provisionend - $provisionstart)) total seconds"
+    echo -e "==> DURATION: $(($provisionend - $provisionstart)) total seconds" | tee -a /catapult/provisioners/redhat/logs/apache.log
 
 elif [ "${4}" = "mysql" ]; then
 
@@ -104,18 +111,21 @@ elif [ "${4}" = "mysql" ]; then
 
     cat "/catapult/provisioners/provisioners.yml" | shyaml get-values-0 redhat.servers.redhat_mysql.modules |
     while read -r -d $'\0' key value; do
-        start=$(date +%s)
-        echo -e "\n\n\n==> MODULE: ${key}"
-        echo -e "==> DESCRIPTION: $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.description)"
-        bash "/catapult/provisioners/redhat/modules/${key}.sh" $1 $2 $3 $4
-        end=$(date +%s)
-        echo -e "==> MODULE: ${key}"
-        echo -e "==> DURATION: $(($end - $start)) seconds"
+        # first boot || dev || not config related || config related and incoming config changes
+        if ([ -s /catapult/provisioners/redhat/logs/mysql.log ]) || ([ $1 == "dev" ]) || ([ $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.configuration) == "False" ]) || ([ $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.configuration) == "True" ] && [ $(cd /catapult && sudo git diff --exit-code ${branch} origin/${branch} "secrets/configuration.yml.gpg") ]); then
+            start=$(date +%s)
+            echo -e "\n\n\n==> MODULE: ${key}"
+            echo -e "==> DESCRIPTION: $(cat "/catapult/provisioners/provisioners.yml" | shyaml get-value redhat.modules.$key.description)"
+            bash "/catapult/provisioners/redhat/modules/${key}.sh" $1 $2 $3 $4
+            end=$(date +%s)
+            echo -e "==> MODULE: ${key}"
+            echo -e "==> DURATION: $(($end - $start)) seconds"
+        fi
     done
 
     provisionend=$(date +%s)
     echo -e "\n\n\n==> PROVISION: mysql"
-    echo -e "==> DURATION: $(($provisionend - $provisionstart)) total seconds"
+    echo -e "==> DURATION: $(($provisionend - $provisionstart)) total seconds" | tee -a /catapult/provisioners/redhat/logs/mysql.log
 
 else
     "Error: Cannot detect the server type."
